#I250: Moved the file from otel to services
